requirements.txt
Flask==3.0.3
flask-cors==6.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.43
google-cloud-secret-manager==2.20.2
requests
pytz
pydantic==2.9.2

# Development tools
pre-commit==4.3.0
ruff==0.13.2
mypy==1.18.2
types-SQLAlchemy==1.4.53.38